Environment variable rework; added env var support for trace id ratio. Unit test refactor / new tests added
